Extract feedback and artifact surface tests
