+++ Tests/QuillCodeAppTests/QuillCodeToolCardSurfaceTests.swift
@@ -1,0 +1,75 @@
+import XCTest
+@testable import QuillCodeApp
+
+final class QuillCodeToolCardSurfaceTests: XCTestCase {
+    func testArtifactStateDerivesLinksAndImagePreviews() {
+        let imageFile = ToolArtifactState(value: "/tmp/quillcode/screenshot.png")
+        XCTAssertEqual(imageFile.kind, .file)
+        XCTAssertEqual(imageFile.href, "file:///tmp/quillcode/screenshot.png")
+        XCTAssertTrue(imageFile.isImagePreview)
+        XCTAssertEqual(imageFile.previewURL, imageFile.href)
+        XCTAssertEqual(imageFile.imagePreview?.typeLabel, "Image")
+        XCTAssertEqual(imageFile.imagePreview?.extensionLabel, "PNG")
+        XCTAssertEqual(imageFile.imagePreview?.detail, "/tmp/quillcode")
+
+        let imageURL = ToolArtifactState(value: "https://example.com/assets/mock.webp?size=large")
+        XCTAssertEqual(imageURL.kind, .url)
+        XCTAssertEqual(imageURL.href, "https://example.com/assets/mock.webp?size=large")
+        XCTAssertEqual(imageURL.label, "example.com/assets/mock.webp")
+        XCTAssertTrue(imageURL.isImagePreview)
+        XCTAssertEqual(imageURL.previewURL, imageURL.href)
+        XCTAssertEqual(imageURL.imagePreview?.extensionLabel, "WEBP")
+        XCTAssertEqual(imageURL.imagePreview?.detail, "example.com/assets/mock.webp")
+
+        let inlineImage = ToolArtifactState(value: "data:image/png;base64,AAAA")
+        XCTAssertEqual(inlineImage.kind, .url)
+        XCTAssertEqual(inlineImage.label, "Inline image")
+        XCTAssertEqual(inlineImage.detail, "Image artifact")
+        XCTAssertTrue(inlineImage.isImagePreview)
+        XCTAssertEqual(inlineImage.previewURL, "data:image/png;base64,AAAA")
+        XCTAssertEqual(inlineImage.imagePreview?.extensionLabel, "PNG")
+        XCTAssertEqual(inlineImage.imagePreview?.detail, "Image artifact")
+        XCTAssertNil(inlineImage.textPreview)
+
+        let nonImageData = ToolArtifactState(value: "data:text/plain;base64,SGVsbG8=")
+        XCTAssertEqual(nonImageData.kind, .path)
+        XCTAssertEqual(nonImageData.label, "data:text/plain;base64,SGVsbG8=")
+        XCTAssertFalse(nonImageData.isImagePreview)
+        XCTAssertNil(nonImageData.previewURL)
+        XCTAssertNil(nonImageData.imagePreview)
+        XCTAssertNil(nonImageData.href)
+        XCTAssertNil(nonImageData.textPreview)
+    }
+
+    func testArtifactStateDerivesDocumentPreviews() {
+        let pdfFile = ToolArtifactState(value: "/tmp/quillcode/reports/briefing.pdf")
+        XCTAssertEqual(pdfFile.kind, .file)
+        XCTAssertFalse(pdfFile.isImagePreview)
+        XCTAssertTrue(pdfFile.isDocumentPreview)
+        XCTAssertEqual(pdfFile.documentPreview?.kind, .pdf)
+        XCTAssertEqual(pdfFile.documentPreview?.typeLabel, "PDF")
+        XCTAssertEqual(pdfFile.documentPreview?.extensionLabel, "PDF")
+        XCTAssertEqual(pdfFile.documentPreview?.detail, "/tmp/quillcode/reports")
+
+        let spreadsheetURL = ToolArtifactState(value: "https://example.com/artifacts/budget.xlsx?download=1")
+        XCTAssertEqual(spreadsheetURL.kind, .url)
+        XCTAssertTrue(spreadsheetURL.isDocumentPreview)
+        XCTAssertEqual(spreadsheetURL.documentPreview?.kind, .spreadsheet)
+        XCTAssertEqual(spreadsheetURL.documentPreview?.typeLabel, "Spreadsheet")
+        XCTAssertEqual(spreadsheetURL.documentPreview?.extensionLabel, "XLSX")
+        XCTAssertEqual(spreadsheetURL.documentPreview?.detail, "example.com/artifacts/budget.xlsx")
+        XCTAssertEqual(spreadsheetURL.href, "https://example.com/artifacts/budget.xlsx?download=1")
+
+        let appshotBundle = ToolArtifactState(value: "/tmp/quillcode/appshots/checkout.appshot.json")
+        XCTAssertEqual(appshotBundle.kind, .file)
+        XCTAssertTrue(appshotBundle.isDocumentPreview)
+        XCTAssertEqual(appshotBundle.documentPreview?.kind, .appshot)
+        XCTAssertEqual(appshotBundle.documentPreview?.typeLabel, "Appshot")
+        XCTAssertEqual(appshotBundle.documentPreview?.extensionLabel, "APPSHOT")
+        XCTAssertEqual(appshotBundle.documentPreview?.detail, "/tmp/quillcode/appshots")
+
+        let textFile = ToolArtifactState(value: "/tmp/quillcode/notes.md", textPreview: "# Notes\n")
+        XCTAssertFalse(textFile.isDocumentPreview)
+        XCTAssertTrue(textFile.hasTextPreview)
+    }
+}
