+++ docs/DECISIONS.md
@@ -262,3 +262,4 @@
 - Browser integration coverage lives in `WorkspaceBrowserIntegrationTests`, not `WorkspaceModelTests`. Browser URL normalization, static and fetched HTML snapshots, fetch-failure fallback, comments, history/reload, invalid-address errors, and composer-driven `host.browser.inspect` cross browser engine state, workspace surfaces, tool cards, and transcript events; keep those flows together while pure URL resolution and browser engine reducers stay in focused unit tests.
 - Workspace review integration coverage lives in `WorkspaceReviewIntegrationTests`, not `WorkspaceModelTests`. Apply-patch review diff refresh, local and SSH Remote stage/restore actions, hunk staging, review comments, tool cards, and review-surface visibility cross the workspace model, git tools, remote SSH executor, transcript events, and review surface; keep those flows together while pure review comment planning and git command construction stay in focused unit tests.
 - Composer integration coverage lives in `WorkspaceComposerIntegrationTests`, not `WorkspaceModelTests`. Composer submit, tool-card creation, artifact surfacing, Computer Use dispatch, queued-tool progress, streaming assistant drafts, cancellation notices, empty-draft no-ops, and selection races cross the workspace model, agent runner, transcript events, tool cards, and top-bar status; keep those flows together while pure submission/cancellation planning stays in focused planner tests.
+- Message feedback and artifact preview coverage should stay out of `WorkspaceModelTests`. Feedback persistence lives in `WorkspaceFeedbackIntegrationTests` because it crosses model mutation, thread events, and transcript surfaces. Pure `ToolArtifactState` image/document/text preview derivation lives in `QuillCodeToolCardSurfaceTests` because it is value-type surface behavior independent of workspace state.
